Add environment variable: ELM_PROTOBUF_EXACTLY_ONEOF
When this environment variable is set to the value "true", the generated
code for the oneof field won't contain the variant "xUnspecified", where
"x" is the name of the oneof.

If this environment variable is unset, or set to anything else, a
variant with the name "xUnspecified" is generated, (the behavior prior
to this commit).